Integrate Mini-OS build with mirage-xen
Instead of installing the various individual object files to ocamlfind,
we now produce a libminios.a with the core kernel.

libminios.a is built using make and the original (almost unmodified)
Mini-OS build system, rather then with ocamlbuild.

The actual OCaml main is missing, so running one of these images just
produces the Mini-OS self-test output.