Updated block_domain for new Mini-OS
Sleeping works differently on ARM, but luckily Mini-OS provides a
block_domain function for us.

Note: select_next's doc comment says it returns "the earliest time
when one sleeping thread will wake up", yet the implementation appeared
to be returning an interval. I changed the implementation to match the
spec, as that seemed saner. This does mean that the time_fun argument
is now unused.